Restructure some files
Made CSInfoFrame a separate file.
Made the single instance of Translator a global constant.
